requirements.txt
a2wsgi==1.10.8
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiologic==0.14.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1732732445543/work
annotated-types @ file:///croot/annotated-types_1709542908624/work
anyio==4.9.0
apache-airflow==3.0.0
apache-airflow-core==3.0.0
apache-airflow-providers-amazon==9.6.1
apache-airflow-providers-common-compat==1.6.0
apache-airflow-providers-common-io==1.5.4
apache-airflow-providers-common-sql==1.26.0
apache-airflow-providers-http==5.2.2
apache-airflow-providers-smtp==2.0.3
apache-airflow-providers-standard==1.0.0
apache-airflow-task-sdk==1.0.0
archspec @ file:///croot/archspec_1709217642129/work
argcomplete==3.6.2
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boltons @ file:///croot/boltons_1737061692168/work
boto3==1.38.1
botocore==1.38.1
Brotli @ file:///croot/brotli-split_1736182456865/work
cadwyn==5.3.3
certifi @ file:///croot/certifi_1738623731865/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorlog==6.9.0
conda @ file:///croot/conda_1738168370816/work
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1736524596117/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1739298990519/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1737733694612/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda-package-streaming @ file:///croot/conda-package-streaming_1731366181659/work
cron-descriptor==1.4.5
croniter==6.0.0
cryptography @ file:///croot/cryptography_1732129644106/work
Deprecated==1.2.18
dill==0.4.0
distro @ file:///croot/distro_1714488253808/work
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==3.1.0
frozendict @ file:///croot/frozendict_1713194832637/work
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.6.1
inflection==0.5.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpath-ng==1.7.0
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
libcst==1.7.0
libmambapy @ file:///croot/mamba-split_1734469461757/work/libmambapy
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.4.0
Mako==1.3.10
markdown-it-py @ file:///work/perseverance-python-buildout/croot/markdown-it-py_1698846045803/work
MarkupSafe==3.0.2
mdurl @ file:///work/perseverance-python-buildout/croot/mdurl_1698845653285/work
menuinst @ file:///croot/menuinst_1738943416351/work
methodtools==0.4.7
more-itertools==10.7.0
msgspec==0.19.0
multidict==6.4.3
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
packaging @ file:///croot/packaging_1734472117206/work
pathspec==0.12.1
pendulum==3.1.0
planetary-computer==1.0.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///croot/pluggy_1733169602837/work
ply==3.11
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg2==2.9.10
PyAthena==3.12.2
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.11.3
pydantic_core==2.33.1
Pygments @ file:///work/perseverance-python-buildout/croot/pygments_1698846270603/work
PyJWT==2.10.1
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
pystac==1.13.0
pystac-client==0.8.6
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-dotenv
python-multipart==0.0.20
python-slugify==8.0.4
python3-saml==1.16.0
pytz==2025.2
PyYAML==6.0.2
redshift-connector==2.1.5
referencing==0.36.2
requests @ file:///croot/requests_1730999120400/work
requests-toolbelt==1.0.0
retryhttp==1.3.2
rich @ file:///croot/rich_1732638981168/work
rich-argparse==1.7.0
rich-toolkit==0.14.3
rpds-py==0.24.0
ruamel.yaml @ file:///croot/ruamel.yaml_1727980156079/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1727769819935/work
s3transfer==0.12.0
sagemaker_studio==1.0.13
scramp==1.4.5
setproctitle==1.3.5
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
starlette==0.46.2
structlog==25.2.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.0.1
text-unidecode==1.3
tqdm @ file:///croot/tqdm_1738943501192/work
truststore @ file:///croot/truststore_1736550121485/work
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3 @ file:///croot/urllib3_1737133630106/work
uuid6==2024.7.10
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
watchtower==3.4.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wirerope==1.0.0
wrapt==1.17.2
xmlsec==1.3.14
yarl==1.20.0
zipp==3.21.0
zstandard @ file:///croot/zstandard_1731356346222/work